Ajusta tamaños de los breadcrumbs

+++ .eleventy.js
@@ -16,7 +16,7 @@ module.exports = function (eleventyConfig) {
 
   eleventyConfig.addFilter("toGLMonthDay", function(value) {
     if (value) {
-      return value.toLocaleDateString("gl", {month: "long", day: "numeric"})
+      return value.toLocaleDateString("gl", {month: "short", day: "numeric"})
     }
     return ""
   });
